int to float64 bugfix
bugfix
"json: cannot unmarshal number 0.6 into Go struct field Part.forecast.parts.prec_mm of type int
exit status 1"
int to float64

+++ wheather/weather.go
@@ -262,7 +262,7 @@ type Part struct {
 	PressureMm int `json:"pressure_mm"` // Давление (в мм рт. ст.)
 	PressurePa int `json:"pressure_pa"` // Давление (в гектопаскалях)
 	Humidity   int `json:"humidity"`    // Влажность воздуха (в процентах)
-	PrecMm     int `json:"prec_mm"`     // Прогнозируемое количество осадков (в мм)
+	PrecMm     float64 `json:"prec_mm"`     // Прогнозируемое количество осадков (в мм)
 	PrecPeriod int `json:"prec_period"` // Прогнозируемый период осадков (в минутах)
 	PrecProb   int `json:"prec_prob"`   // Вероятность выпадения осадков
 }
